refactor architecture specific code
Before, the architecture specific code was layed
out in files matching different architectures.
This lead to code duplication of e.g. OpenSSL secure memory
handling in different places.

Therefore this commit reorganizes the architecture specific code
around common abstractions, which allows for better code reuse.

Furthermore it adds support for more platforms, like musl libc.
There jent_ncpu always returned 1 before.
