--- arch/jitterentropy-base-s390.h
@@ -1,185 +1,0 @@
-/*
- * Non-physical true random number generator based on timing jitter.
- *
- * Copyright Stephan Mueller <smueller@chronox.de>, 2013 - 2026
- *
- * License
- * =======
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- * 1. Redistributions of source code must retain the above copyright
- *    notice, and the entire permission notice in its entirety,
- *    including the disclaimer of warranties.
- * 2. Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in the
- *    documentation and/or other materials provided with the distribution.
- * 3. The name of the author may not be used to endorse or promote
- *    products derived from this software without specific prior
- *    written permission.
- *
- * ALTERNATIVELY, this product may be distributed under the terms of
- * the GNU General Public License, in which case the provisions of the GPL are
- * required INSTEAD OF the above restrictions.  (This clause is
- * necessary due to a potential bad interaction between the GPL and
- * the restrictions contained in a BSD-style copyright.)
- *
- * THIS SOFTWARE IS PROVIDED ``AS IS'' AND ANY EXPRESS OR IMPLIED
- * WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
- * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE, ALL OF
- * WHICH ARE HEREBY DISCLAIMED.  IN NO EVENT SHALL THE AUTHOR BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT
- * OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR
- * BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
- * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
- * USE OF THIS SOFTWARE, EVEN IF NOT ADVISED OF THE POSSIBILITY OF SUCH
- * DAMAGE.
- */
-
-#ifndef _JITTERENTROPY_BASE_S390_H
-#define _JITTERENTROPY_BASE_S390_H
-
-#include <sched.h>
-#include <stdint.h>
-#include <stdlib.h>
-#include <string.h>
-
-static inline void jent_get_nstime(uint64_t *out)
-{
-	/*
-	 * This is MVS+STCK code! Enable it with -S in the compiler.
-	 *
-	 * uint64_t clk;
-	 * __asm__ volatile("stck %0" : "=m" (clk) : : "cc");
-	 * *out = (uint64_t)(clk);
-	 */
-
-	/*
-	 * This is GCC+STCKE code. STCKE command and data format:
-	 * z/Architecture - Principles of Operation
-	 * http://publibz.boulder.ibm.com/epubs/pdf/dz9zr007.pdf
-	 *
-	 * The current value of bits 0-103 of the TOD clock is stored in bytes
-	 * 1-13 of the sixteen-byte output:
-	 *
-	 * bits 0-7: zeros (reserved for future extention)
-	 * bits 8-111: TOD Clock value
-	 * bits 112-127: Programmable Field
-	 *
-	 * Output bit 59 (TOD-Clock bit 51) effectively increments every
-	 * microsecond. Bits 60 to 111 of STCKE output are fractions of
-	 * a miscrosecond: bit 59 is 1.0us, bit 60 is .5us, bit 61 is .25us,
-	 * bit 62 is .125us, bit 63 is 62.5ns, etc.
-	 *
-	 * Some of these bits can be implemented, some not. 64 bits of
-	 * the TOD clock are implemented usually nowadays, these are
-	 * bits 8-71 of the output.
-	 *
-	 * The stepping value of TOD-clock bit position 63, if implemented,
-	 * is 2^-12 microseconds, or approximately 244 picoseconds. This value
-	 * is called a clock unit.
-	 */
-
-	uint8_t clk[16];
-
-        asm volatile("stcke %0" : "=Q" (clk) : : "cc");
-
-	/* s390x is big-endian, so just perfom a byte-by-byte copy */
-	*out = *(uint64_t *)(clk + 1);
-}
-
-static inline void *jent_zalloc(size_t len)
-{
-	void *tmp = NULL;
-	/* we have no secure memory allocation! Hence
-	 * we do not set CRYPTO_CPU_JITTERENTROPY_SECURE_MEMORY */
-	tmp = malloc(len);
-	if(NULL != tmp)
-		memset(tmp, 0, len);
-	return tmp;
-}
-
-static inline void jent_memset_secure(void *s, size_t n)
-{
-	memset(s, 0, n);
-	__asm__ __volatile__("" : : "r" (s) : "memory");
-}
-
-static inline void jent_zfree(void *ptr, size_t len)
-{
-	jent_memset_secure(ptr, len);
-	free(ptr);
-}
-
-static inline int jent_fips_enabled(void)
-{
-        return 0;
-}
-
-static inline long jent_ncpu(void)
-{
-	/*
-	 * TODO: return number of available CPUs -
-	 * this code disables timer thread as only one CPU is "detected".
-	 */
-	return 1;
-}
-
-static inline void jent_yield(void)
-{
-	sched_yield();
-}
-
-static inline uint32_t jent_cache_size_roundup(int all_caches)
-{
-#ifdef __linux__
-	long l1 = sysconf(_SC_LEVEL1_DCACHE_SIZE);
-	uint32_t cache_size = 0;
-
-	/* Cache size reported by system */
-	if (l1 > 0)
-		cache_size += (uint32_t)l1;
-
-	if (all_caches) {
-		long l2 = sysconf(_SC_LEVEL2_CACHE_SIZE);
-		long l3 = sysconf(_SC_LEVEL3_CACHE_SIZE);
-
-		if (l2 > 0)
-			cache_size += (uint32_t)l2;
-		if (l3 > 0)
-			cache_size += (uint32_t)l3;
-	}
-
-	/* Force the output_size to be of the form (bounding_power_of_2 - 1). */
-	cache_size |= (cache_size >> 1);
-	cache_size |= (cache_size >> 2);
-	cache_size |= (cache_size >> 4);
-	cache_size |= (cache_size >> 8);
-	cache_size |= (cache_size >> 16);
-
-	if (cache_size == 0)
-		return 0;
-
-	/* Make the output_size the smallest power of 2 strictly greater than cache_size. */
-	cache_size++;
-
-	return cache_size;
-#else
-	(void)all_caches;
-	return 0;
-#endif
-}
-
-/* --- helpers needed in user space -- */
-
-static inline uint64_t rol64(uint64_t x, int n)
-{
-	return ( (x << (n&(64-1))) | (x >> ((64-n)&(64-1))) );
-}
-
-
-#endif /* _JITTERENTROPY_BASE_S390_H */
-
