+++ src/jitterentropy-sha3.c
@@ -54,6 +54,11 @@ static inline void le64_to_ptr(uint8_t *p, const uint64_t value)
 	le32_to_ptr(p,     (uint32_t)(value));
 }
 
+static inline uint64_t rol64(uint64_t x, int n)
+{
+	return ( (x << (n&(64-1))) | (x >> ((64-n)&(64-1))) );
+}
+
 /*********************************** Keccak ***********************************/
 /* state[x + y*5] */
 #define A(x, y) (x + 5 * y)
